deb: fix processing several fields from a Release file
Before the patch, the following expression was used when processing
the "Architectures" and "Components" fields from the release file:
`architectures = release['Architectures'].split(' ')`
But in the case of `release['Architectures'] == ''` `architectures`
will be `['']` and the following `for` will not work correctly
(it will iterate for an empty `arch`).
Now in the case of `release['Architectures'] == ''` `architectures`
will be `[]` and `for` would not iterate over a such list.